misc: rename stylesheets
