+++ src/components/common/SegmentedInput/index.module.scss
