+++ src/components/game/Icon/icon.module.scss
