+++ src/components/game/Score/score.module.scss
