+++ src/components/gameEnd/Table/table.module.scss
