+++ src/components/layout/Footer/index.jsx
@@ -1,4 +1,4 @@
-import styles from './index.module.scss'
+import styles from './footer.module.scss'
 
 const Footer = ({ children }) => {
 	return <footer className={styles.footer}>{children}</footer>
